+++ gtk/gtknative.h
@@ -49,8 +49,8 @@ struct _GtkNativeInterface
   GskRenderer * (* get_renderer)          (GtkNative    *self);
 
   void          (* get_surface_transform) (GtkNative    *self,
-                                           int          *x,
-                                           int          *y);
+                                           double       *x,
+                                           double       *y);
 
   void          (* check_resize)          (GtkNative    *self);
 };
@@ -69,8 +69,8 @@ GskRenderer *gtk_native_get_renderer   (GtkNative *self);
 
 GDK_AVAILABLE_IN_ALL
 void         gtk_native_get_surface_transform (GtkNative *self,
-                                               int       *x,
-                                               int       *y);
+                                               double    *x,
+                                               double    *y);
 
 G_END_DECLS
